requirements.txt
Flask==2.1.2
requests==2.26.0
openai==0.10.5
python-dotenv==0.19.1